+++ mex/editor/edit/models.py
@@ -7,10 +7,21 @@ from mex.editor.models import EditorValue
 class InputConfig(rx.Base):
     """Model for configuring input masks."""
 
-    badge_options: list[str] = []
-    editable_href: bool = False
-    editable_badge: bool = False
-    editable_text: bool = False
+    badge_default: str | None = None  # value to pre-select in drop-down menu
+    badge_options: list[str] = []  # possible values to show in drop-down menu
+    badge_titles: list[str] = []  # title for the collection of drop-drown choices
+    editable_href: bool = False  # whether the href attribute is editable as text
+    editable_badge: bool = False  # whether the badge is editable as a drop-down
+    editable_text: bool = False  # whether the text is editable as plain text
+    allow_additive: bool = False  # whether this field belongs to an additive rule
+
+
+class ValidationMessage(rx.Base):
+    """Model for describing validation errors."""
+
+    field_name: str
+    message: str
+    input: str
 
 
 class EditorPrimarySource(rx.Base):
@@ -18,7 +29,7 @@ class EditorPrimarySource(rx.Base):
 
     name: EditorValue
     identifier: MergedPrimarySourceIdentifier
-    input_config: InputConfig | None
+    input_config: InputConfig
     editor_values: list[EditorValue] = []
     enabled: bool = True