+++ mex/editor/edit/state.py
@@ -1,6 +1,7 @@
 from collections.abc import Generator
 
 import reflex as rx
+from pydantic import ValidationError
 from reflex.event import EventSpec
 from requests import HTTPError
 from starlette import status
@@ -8,10 +9,11 @@ from starlette import status
 from mex.common.backend_api.connector import BackendApiConnector
 from mex.common.models import RULE_SET_RESPONSE_CLASSES_BY_NAME
 from mex.common.transform import ensure_postfix
-from mex.editor.edit.models import EditorField, EditorPrimarySource
+from mex.editor.edit.models import EditorField, EditorPrimarySource, ValidationMessage
 from mex.editor.edit.transform import (
     transform_fields_to_rule_set,
     transform_models_to_fields,
+    transform_validation_error_to_messages,
 )
 from mex.editor.exceptions import escalate_error
 from mex.editor.models import EditorValue
@@ -29,6 +31,7 @@ class EditState(State):
     fields: list[EditorField] = []
     item_title: list[EditorValue] = []
     stem_type: str | None = None
+    validation_messages: list[ValidationMessage] = []
 
     @rx.event(background=True)
     async def resolve_identifiers(self):
@@ -97,7 +100,11 @@ class EditState(State):
         if self.stem_type is None:
             self.reset()
             return
-        rule_set = transform_fields_to_rule_set(self.stem_type, self.fields)
+        try:
+            rule_set = transform_fields_to_rule_set(self.stem_type, self.fields)
+        except ValidationError as exc:
+            self.validation_messages = transform_validation_error_to_messages(exc)
+            return
         connector = BackendApiConnector.get()
         try:
             # TODO(ND): use proper connector method when available (stop-gap MX-1762)
@@ -116,7 +123,7 @@ class EditState(State):
         resolve_identifier.cache_clear()
         yield rx.toast.success(
             title="Saved",
-            description=f"{self.stem_type} rule-set was saved successfully.",
+            description=f"{self.stem_type} was saved successfully.",
             class_name="editor-toast",
             close_button=True,
             dismissible=True,
@@ -124,13 +131,29 @@ class EditState(State):
         )
         yield from self.refresh()
 
+    @rx.event
+    def clear_validation_messages(self) -> None:
+        """Clear all validation messages."""
+        self.validation_messages.clear()
+
     def _get_primary_sources_by_field_name(
         self, field_name: str
     ) -> list[EditorPrimarySource]:
+        """Get all primary sources for the given field name."""
         for field in self.fields:
             if field.name == field_name:
                 return field.primary_sources
         msg = f"field not found: {field_name}"
+        raise ValueError(msg)
+
+    def _get_editable_primary_source_by_field_name(
+        self, field_name: str
+    ) -> EditorPrimarySource:
+        """Get the (first) primary source that allows an editable rule."""
+        for primary_source in self._get_primary_sources_by_field_name(field_name):
+            if primary_source.input_config.allow_additive:
+                return primary_source
+        msg = f"editable field not found: {field_name}"
         raise ValueError(msg)
 
     @rx.event
@@ -161,35 +184,30 @@ class EditState(State):
     @rx.event
     def add_additive_value(self, field_name: str) -> None:
         """Add an additive rule to the given field."""
-        for primary_source in self._get_primary_sources_by_field_name(field_name):
-            if primary_source.input_config:
-                primary_source.editor_values.append(EditorValue())
+        primary_source = self._get_editable_primary_source_by_field_name(field_name)
+        primary_source.editor_values.append(EditorValue())
 
     @rx.event
     def remove_additive_value(self, field_name: str, index: int) -> None:
         """Remove an additive rule from the given field."""
-        for primary_source in self._get_primary_sources_by_field_name(field_name):
-            if primary_source.input_config:
-                primary_source.editor_values.pop(index)
+        primary_source = self._get_editable_primary_source_by_field_name(field_name)
+        primary_source.editor_values.pop(index)
 
     @rx.event
     def set_text_value(self, field_name: str, index: int, value: str) -> None:
         """Set the text attribute on an additive editor value."""
-        for primary_source in self._get_primary_sources_by_field_name(field_name):
-            if primary_source.input_config:
-                primary_source.editor_values[index].text = value
+        primary_source = self._get_editable_primary_source_by_field_name(field_name)
+        primary_source.editor_values[index].text = value
 
     @rx.event
     def set_badge_value(self, field_name: str, index: int, value: str) -> None:
         """Set the badge attribute on an additive editor value."""
-        for primary_source in self._get_primary_sources_by_field_name(field_name):
-            if primary_source.input_config:
-                primary_source.editor_values[index].badge = value
+        primary_source = self._get_editable_primary_source_by_field_name(field_name)
+        primary_source.editor_values[index].badge = value
 
     @rx.event
     def set_href_value(self, field_name: str, index: int, value: str) -> None:
         """Set an external href on an additive editor value."""
-        for primary_source in self._get_primary_sources_by_field_name(field_name):
-            if primary_source.input_config:
-                primary_source.editor_values[index].href = value
-                primary_source.editor_values[index].external = True
+        primary_source = self._get_editable_primary_source_by_field_name(field_name)
+        primary_source.editor_values[index].href = value
+        primary_source.editor_values[index].external = True