+++ tests/test_transform.py
@@ -1,6 +1,5 @@
 import pytest
 
-from mex.common.exceptions import MExError
 from mex.common.models import AnyExtractedModel
 from mex.common.types import APIType, Identifier, Link, LinkLanguage, Text, TextLanguage
 from mex.editor.models import EditorValue
@@ -32,16 +31,10 @@ from mex.editor.transform import (
             True,
             [
                 EditorValue(text="bar"),
-                EditorValue(text="REST", badge="APIType"),
+                EditorValue(text="APIType", badge="REST"),
+                EditorValue(text="hi there", badge="EN"),
                 EditorValue(
-                    text="hi there",
-                    badge="en",
-                ),
-                EditorValue(
-                    text="homepage",
-                    badge="en",
-                    href="http://mex",
-                    external=True,
+                    text="homepage", badge="EN", href="http://mex", external=True
                 ),
             ],
         ),
@@ -73,7 +66,9 @@ def test_transform_values(
 
 
 def test_transform_value_none_error() -> None:
-    with pytest.raises(MExError, match="cannot transform NoneType to editor value"):
+    with pytest.raises(
+        NotImplementedError, match="cannot transform NoneType to editor value"
+    ):
         transform_value(None)
 
 
@@ -94,29 +89,19 @@ def test_transform_models_to_title(dummy_data: list[AnyExtractedModel]) -> None:
     assert dummy_titles == [
         [
             # ps-1 primary source renders title as text
-            EditorValue(
-                text="Primary Source One",
-                badge="en",
-            )
+            EditorValue(text="Primary Source One", badge="EN")
         ],
         [
             # ps-2 primary source renders title as text
-            EditorValue(
-                text="Primary Source Two",
-                badge="en",
-            )
+            EditorValue(text="Primary Source Two", badge="EN")
         ],
         [
             # contact-point renders email as text
-            EditorValue(
-                text="info@contact-point.one",
-            )
+            EditorValue(text="info@contact-point.one")
         ],
         [
             # contact-point renders email as text
-            EditorValue(
-                text="help@contact-point.two",
-            )
+            EditorValue(text="help@contact-point.two")
         ],
         [
             # unit renders shortName as text (no language badge)
@@ -124,10 +109,11 @@ def test_transform_models_to_title(dummy_data: list[AnyExtractedModel]) -> None:
         ],
         [
             # activity renders title as text (with language badge)
-            EditorValue(
-                text="Aktivität 1",
-                badge="de",
-            )
+            EditorValue(text="Aktivität 1", badge="DE")
+        ],
+        [
+            # resource renders title as text
+            EditorValue(text="Bioinformatics Resource 1"),
         ],
     ]
 
@@ -143,34 +129,19 @@ def test_transform_models_to_preview(dummy_data: list[AnyExtractedModel]) -> Non
         [EditorValue(text="PrimarySource")],
         [EditorValue(text="ContactPoint")],
         [EditorValue(text="ContactPoint")],
-        [
-            EditorValue(
-                text="Unit 1",
-                badge="en",
-                enabled=True,
-            )
-        ],
+        [EditorValue(text="Unit 1", badge="EN", enabled=True)],
         [
             EditorValue(text="A1", enabled=True),
-            EditorValue(
-                identifier="wEvxYRPlmGVQCbZx9GAbn",
-            ),
-            EditorValue(
-                identifier="g32qzYNVH1Ez7JTEk3fvLF",
-            ),
-            EditorValue(
-                identifier="cWWm02l1c6cucKjIhkFqY4",
-            ),
-            EditorValue(
-                identifier="cWWm02l1c6cucKjIhkFqY4",
-            ),
-            EditorValue(
-                text="1999-12-24",
-                badge="day",
-            ),
-            EditorValue(
-                text="2023-01-01",
-                badge="day",
-            ),
+            EditorValue(identifier="wEvxYRPlmGVQCbZx9GAbn"),
+            EditorValue(identifier="g32qzYNVH1Ez7JTEk3fvLF"),
+            EditorValue(identifier="cWWm02l1c6cucKjIhkFqY4"),
+            EditorValue(identifier="cWWm02l1c6cucKjIhkFqY4"),
+            EditorValue(text="1999-12-24", badge="day"),
+            EditorValue(text="2023-01-01", badge="day"),
+        ],
+        [
+            EditorValue(identifier="cWWm02l1c6cucKjIhkFqY4"),
+            EditorValue(text="Theme", badge="BIOINFORMATICS_AND_SYSTEMS_BIOLOGY"),
+            EditorValue(text="AccessRestriction", badge="OPEN"),
         ],
     ]